fix(cycle): working module-settings button + align year selector
- The « Paramètre du module » button did nothing: useModuleSettings was
  called in the page component, which renders ModuleShell and so sits
  ABOVE its ModuleSettingsProvider (returns null). Move the settings
  panel into a child (CycleSettingsPanel) mounted under the provider,
  like Mood's PrimaryColumn.
- Vertically centre the year selector with the module title (items-center,
  drop the -mt-1).
