+++ packages/web/src/app/flow/Cycle/components/CycleViews.tsx
@@ -92,7 +92,7 @@ export default function CycleViews({
 
   return (
     <div className="sticky top-13 z-10 -mt-7 bg-bg pt-7 pb-3">
-      <div className="flex flex-wrap items-start justify-between gap-x-6 gap-y-2">
+      <div className="flex flex-wrap items-center justify-between gap-x-6 gap-y-2">
         <PageHeading className="mb-0 hidden lg:block">{t('cycle.title')}</PageHeading>
         <CycleYearSelector
           year={year}
